Implement/delete app

+++ src/app-auth/services/app-auth.service.ts
@@ -7,7 +7,7 @@ import {
   BadRequestException,
 } from '@nestjs/common';
 import { SigningStargateClient } from '@cosmjs/stargate';
-import { CreateAppDto } from '../dtos/create-app.dto';
+import { CreateAppDto, DeleteAppResponse } from '../dtos/create-app.dto';
 import { App, createAppResponse } from 'src/app-auth/schemas/app.schema';
 import { AppRepository } from '../repositories/app.repository';
 import { UpdateAppDto } from '../dtos/update-app.dto';
@@ -35,6 +35,8 @@ import {
   MSG_UPDATE_DID_TYPEURL,
 } from 'src/utils/authz';
 import { AuthzCreditService } from 'src/credits/services/credits.service';
+import { AuthZCreditsRepository } from 'src/credits/repositories/authz.repository';
+import { EdvClientKeysManager } from 'src/edv/services/edv.singleton';
 
 enum GRANT_TYPES {
   access_service_kyc = 'access_service_kyc',
@@ -56,6 +58,7 @@ export class AppAuthService {
     private readonly supportedServices: SupportedServiceService,
     private readonly userRepository: UserRepository,
     private readonly authzCreditService: AuthzCreditService,
+    private readonly authzCreditRepository: AuthZCreditsRepository,
   ) {}
 
   async createAnApp(
@@ -531,7 +534,7 @@ export class AppAuthService {
     return this.getAppResponse(app);
   }
 
-  async deleteApp(appId: string, userId: string): Promise<App> {
+  async deleteApp(appId: string, userId: string): Promise<DeleteAppResponse> {
     Logger.log('deleteApp() method: starts....', 'AppAuthService');
 
     let appDetail = await this.appRepository.findOne({ appId, userId });
@@ -540,13 +543,40 @@ export class AppAuthService {
 
       throw new NotFoundException([`No App found for appId ${appId}`]);
     }
-    //commenting this code as delete operation is not implemented in edvClient
-
-    // const { edvId, edvDocId } = appDetail;
-    // await this.edvService.init(edvId);
-    // await this.edvService.deleteDoc(edvDocId);
+    const { edvId, kmsId } = appDetail;
+    const appDataFromVault = await globalThis.kmsVault.getDecryptedDocument(
+      kmsId,
+    );
+    if (!appDataFromVault) {
+      throw new BadRequestException('App detail does not exists in datavault');
+    }
+    const appKmsVaultWallet = await VaultWalletManager.getWallet(
+      appDataFromVault.mnemonic,
+    );
+    const kmsVaultManager = new EdvClientKeysManager();
+    const appKmsVault = await kmsVaultManager.createVault(
+      appKmsVaultWallet,
+      edvId,
+    );
+    try {
+      await appKmsVault.deleteVault(edvId);
+      await globalThis.kmsVault.deleteDocument(kmsId);
+    } catch (vaultError) {
+      Logger.error(
+        `Error deleting KMS or EDV vault: ${vaultError}`,
+        'AppAuthService',
+      );
+      throw new BadRequestException(['Failed to delete vault']);
+    }
+    // delete app db also
+    if (!appDetail.services || appDetail.services.length === 0) {
+      throw new BadRequestException(['Invalid app']);
+    }
+    const appDbConnectionSuffix = `service:${appDetail.services[0].dBSuffix}:${appDetail.subdomain}`;
+    await this.appRepository.findAndDeleteServiceDB(appDbConnectionSuffix);
+    this.authzCreditRepository.deleteAuthzDetail({ appId });
     appDetail = await this.appRepository.findOneAndDelete({ appId, userId });
-    return appDetail;
+    return { appId: appDetail.appId };
   }
 
   private checkIfDateExpired(expiryDate: Date | null) {
